renaming: use dash case instead of camelCase
