+++ src/packages/styled-components/__tests__/render-stylesheet.js
