+++ src/packages/styled-components/index.js
@@ -1,1 +1,1 @@
-export { createStyled as default } from './createStyledComponent'
+export { createStyled as default } from './create-styled-component'
